% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/christmas_stats_public-data.R
\docType{data}
\name{christmas_stats_samples}
\alias{christmas_stats_samples}
\title{Cleaned, event-level data}
\description{
Event-level data for the Christmas statistical cognition
experiment
}
\details{
\itemize{
\item id: participant id
\item ev: (absolute value of) strength of "evidence" in experiment.
This is the transformed two-tailed p value.
\item sgn: sign of evidence (-1 means favoring left side)
\item n_idx: index of sample size selected by participant (0-19)
\item n: "true" sample size per group, hidden from the participant
\item d: "true" effect size, hidden from the participant (same as effect_size in christmas_stats_participants)
\item time: javascript timestamp (ms) reported by browser. This was when the point
was place on the interface. Indexed to initial_timestamp
\item toy_name: name of toy randomly assigned to participant
\item initial_timestamp: (should be 0, start of experiment)
\item type: 'expt' for experimental samples, 'null' for
samples from random shuffle reports
\item initiated_timestamp: javascript timestamp (ms) when the sample request was
initiated by the participant. Indexed to initial_timestamp
}
}
\author{
Richard D. Morey \email{richarddmorey@gmail.com}
}
\keyword{data}
