% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_complete_dataset.R
\name{compile_complete_dataset}
\alias{compile_complete_dataset}
\title{Complete data set after extracting samples and making some columns "neater"}
\usage{
compile_complete_dataset()
}
\value{
List containing two data frames:
compiled_data (the participant summaries) and samples (participants'
clicks in the experimental task)
}
\description{
Complete data set after extracting samples and making some columns "neater"
}
