% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_data.R
\name{explore_data}
\alias{explore_data}
\title{Run shiny app to explore the (cleaned) data set}
\usage{
explore_data()
}
\value{
Returns the result of shiny::runApp()
}
\description{
Run shiny app to explore the (cleaned) data set
}
