% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_desc_evidence.R
\name{fig_desc_evidence_transform}
\alias{fig_desc_evidence_transform}
\title{Show the transformation between the z/p value and the visual location}
\usage{
fig_desc_evidence_transform()
}
\value{
Invisibly returns NULL
}
\description{
Show the transformation between the z/p value and the visual location
}
\examples{
fig_desc_evidence_transform()
}
