% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_error_rates.R
\name{fig_dprime_effect_size}
\alias{fig_dprime_effect_size}
\title{d prime as a function of effect size (model fits)}
\usage{
fig_dprime_effect_size(
  filter_func = function(data, ...) return(data),
  print_plot = TRUE
)
}
\arguments{
\item{filter_func}{function to filter the data before plotting}

\item{print_plot}{(logical) set to FALSE to prevent printing the ggplot object}
}
\value{
Invisibly returns the ggplot object
}
\description{
d prime as a function of effect size (model fits)
}
\examples{
suppressWarnings(
  fig_dprime_effect_size()
)
}
