% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_edu_training.R
\name{fig_edu_training_edu}
\alias{fig_edu_training_edu}
\title{Distribution of attained educational level in the sample}
\usage{
fig_edu_training_edu(
  filter_func = function(data, ...) return(data),
  print_plot = TRUE
)
}
\arguments{
\item{filter_func}{function to filter the data before plotting}

\item{print_plot}{(logical) set to FALSE to prevent printing the ggplot object}
}
\value{
Invisibly returns the ggplot object
}
\description{
Distribution of attained educational level in the sample
}
\examples{
suppressWarnings(
  fig_edu_training_edu()
)
}
