% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_edu_training.R
\name{fig_edu_training_stats}
\alias{fig_edu_training_stats}
\title{Distribution of reported number of years of statistical training in the sample}
\usage{
fig_edu_training_stats(filter_func = function(data, ...) return(data),
  print_plot = TRUE)
}
\arguments{
\item{filter_func}{function to filter the data before plotting}

\item{print_plot}{(logical) set to FALSE to prevent printing the ggplot object}
}
\value{
Invisibly returns the ggplot object
}
\description{
Distribution of reported number of years of statistical training in the sample
}
\examples{
fig_edu_training_stats()
}
