% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_error_rates.R
\name{fig_error_rates}
\alias{fig_error_rates}
\title{Error/correct response rates as a function of effect size}
\usage{
fig_error_rates(
  filter_func = function(data, ...) return(data),
  print_plot = TRUE
)
}
\arguments{
\item{filter_func}{function to filter the data before plotting}

\item{print_plot}{(logical) set to FALSE to prevent printing the ggplot object}
}
\value{
Invisibly returns the ggplot object
}
\description{
Error/correct response rates as a function of effect size
}
\examples{
suppressWarnings(
  fig_error_rates()
)
}
