% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_evidence_p_vals.R
\name{fig_evidence_p_vals}
\alias{fig_evidence_p_vals}
\title{Generate image to show the interface with p values overlaid}
\usage{
fig_evidence_p_vals(
  q,
  ps = c(0.005, 0.025, 0.05, 0.125, 0.25, 0.5, 0.75, 0.875, 0.95, 0.975, 0.995),
  letter = ""
)
}
\arguments{
\item{q}{Evidence scale/power}

\item{ps}{vector of p values to plot}

\item{letter}{label to add to plot}
}
\value{
Returns a file path to the generated PNG file
}
\description{
Generate image to show the interface with p values overlaid
}
\examples{
fig_evidence_p_vals(q = 7)
}
