% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_evidence_scale.R
\name{fig_evidence_scale_logistic_p}
\alias{fig_evidence_scale_logistic_p}
\title{Probability of making a "difference" decision as a function of the most extreme p value and evidence scale}
\usage{
fig_evidence_scale_logistic_p(
  pval = 0.05,
  seed = 1,
  shift_y = 0.2,
  around_y = 0.1,
  jitter_y = around_y/4,
  filter_func = function(data, ...) return(data)
)
}
\arguments{
\item{pval}{vector of p values at which to indicate the scale of the respective null sampling distribution}

\item{seed}{random seed for jitter of points}

\item{shift_y}{How far above/below the main graph to put the middle of the "scatter" plots}

\item{around_y}{How far above/below the middle of the "scatter" plots to put the points}

\item{jitter_y}{How much to jutter the points in the y direction}

\item{filter_func}{function to filter the data before plotting}
}
\value{
Invisibly returns the corresponding GLM test result
}
\description{
Probability of making a "difference" decision as a function of the most extreme p value and evidence scale
}
\examples{
fig_evidence_scale_logistic_p()
}
