% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_field.R
\name{fig_field}
\alias{fig_field}
\title{Distribution of self-reported scientific field in the sample}
\usage{
fig_field(filter_func = function(data, ...) return(data),
  print_plot = TRUE)
}
\arguments{
\item{filter_func}{function to filter the data before plotting}

\item{print_plot}{(logical) set to FALSE to prevent printing the ggplot object}
}
\value{
Invisibly returns the ggplot object
}
\description{
Distribution of self-reported scientific field in the sample
}
\examples{
fig_field()
}
