% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_field_scientific.R
\name{fig_field_scientific}
\alias{fig_field_scientific}
\title{Participants' reports of whether others would call their field scientific}
\usage{
fig_field_scientific(
  filter_func = function(data, ...) return(data),
  print_plot = TRUE
)
}
\arguments{
\item{filter_func}{function to filter the data before plotting}

\item{print_plot}{(logical) set to FALSE to prevent printing the ggplot object}
}
\value{
Invisibly returns the ggplot object
}
\description{
Participants' reports of whether others would call their field scientific
}
\examples{
suppressWarnings(
  fig_field_scientific()
)
}
