% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_how_use_stat.R
\name{fig_how_use_stat}
\alias{fig_how_use_stat}
\title{Distribution of self-reported ways participants use statistics}
\usage{
fig_how_use_stat(filter_func = function(data, ...) return(data),
  print_plot = TRUE)
}
\arguments{
\item{filter_func}{function to filter the data before plotting}

\item{print_plot}{(logical) set to FALSE to prevent printing the ggplot object}
}
\value{
Invisibly returns the ggplot object
}
\description{
Distribution of self-reported ways participants use statistics
}
\examples{
fig_how_use_stat()
}
