% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_preferred_method.R
\name{fig_preferred_method}
\alias{fig_preferred_method}
\title{Distribution of self-reported preferences for statistical methods}
\usage{
fig_preferred_method(
  filter_func = function(data, ...) return(data),
  print_plot = TRUE
)
}
\arguments{
\item{filter_func}{function to filter the data before plotting}

\item{print_plot}{(logical) set to FALSE to prevent printing the ggplot object}
}
\value{
Invisibly returns the ggplot object
}
\description{
Distribution of self-reported preferences for statistical methods
}
\examples{
suppressWarnings(
  fig_preferred_method()
)
}
