% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_recreate_display.R
\name{fig_recreate_display}
\alias{fig_recreate_display}
\title{Generate image to show a participant's sampling behaviour}
\usage{
fig_recreate_display(id, type = "expt", ps = c(0.005, 0.025, 0.05,
  0.125, 0.25, 0.5, 0.75, 0.875, 0.95, 0.975, 0.995), letter = "")
}
\arguments{
\item{id}{ID of participant whose samples to show}

\item{type}{String indicating type of samples ("expt" for experiments; "null" for random shuffle reports)}

\item{ps}{p values to overlay on the display}
}
\value{
Returns a file path to the generated PNG file
}
\description{
Generate image to show a participant's sampling behaviour
}
\examples{
fig_recreate_display(id = "R_7V3ufsG6NYL27Ul")
}
