% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_response_evidence.R
\name{fig_response_evidence}
\alias{fig_response_evidence}
\title{Distribution of "evidence" in the display by response}
\usage{
fig_response_evidence(filter_func = function(data, ...) return(data))
}
\arguments{
\item{filter_func}{function to filter the data before plotting}
}
\value{
Invisibly returns NULL
}
\description{
Distribution of "evidence" in the display by response
}
\examples{
fig_response_evidence()
}
