% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_sampling_behavior.R
\name{fig_sampling_behavior}
\alias{fig_sampling_behavior}
\title{Number of samples per participant for each activity, by effect size}
\usage{
fig_sampling_behavior(filter_func = function(data, ...) return(data),
  print_plot = TRUE)
}
\arguments{
\item{filter_func}{function to filter the data before plotting}

\item{print_plot}{(logical) set to FALSE to prevent printing the ggplot object}
}
\value{
Invisibly returns a list containing the ggplot objects
and the result of the robust regression analysis
}
\description{
Number of samples per participant for each activity, by effect size
}
\examples{
fig_sampling_behavior()
}
