% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_shuffle_understanding.R
\name{fig_shuffle_understanding}
\alias{fig_shuffle_understanding}
\title{Distribution of self-reported understanding of what usefulness of random shuffle reports before experiment}
\usage{
fig_shuffle_understanding(
  filter_func = function(data, ...) return(data),
  print_plot = TRUE
)
}
\arguments{
\item{filter_func}{function to filter the data before plotting}

\item{print_plot}{(logical) set to FALSE to prevent printing the ggplot object}
}
\value{
Invisibly returns the ggplot object
}
\description{
Distribution of self-reported understanding of what usefulness of random shuffle reports before experiment
}
\examples{
suppressWarnings(
  fig_shuffle_understanding()
)
}
