% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_strategies.R
\name{fig_strategies}
\alias{fig_strategies}
\title{Frequencies of self-reported strategies}
\usage{
fig_strategies(
  filter_func = function(data, ...) return(data),
  print_plot = TRUE
)
}
\arguments{
\item{filter_func}{function to filter the data before plotting}

\item{print_plot}{(logical) set to FALSE to prevent printing the ggplot object}
}
\value{
Invisibly returns the ggplot object
}
\description{
Coded responses to the three open-ended text questions
about participants' strategies.
}
\examples{
suppressWarnings(
  fig_strategies()
)
}
