% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_clean_dataset.R
\name{load_clean_dataset}
\alias{load_clean_dataset}
\title{Load the data set and apply all cleaning steps}
\usage{
load_clean_dataset()
}
\value{
Returns a list of four objects:
\itemize{
\item clean_data: data frame with cleaned data, per participant
\item samples: Each click for each participant in the experimental task
\item eliminations: number of participants that fail to meet each criterion
\item original rows: the number of rows in the data set before cleaning
}
}
\description{
Load the data set and apply all cleaning steps
}
