% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_complete_dataset.R
\name{load_complete_dataset}
\alias{load_complete_dataset}
\title{Read the complete data set from the (compressed) CSV}
\usage{
load_complete_dataset()
}
\value{
Returns a data from containing the entire data set,
with invalid rows (mobile users and non-consents) filtered out
}
\description{
Read the complete data set from the (compressed) CSV
}
