% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_text_responses.R
\name{load_text_responses}
\alias{load_text_responses}
\title{Text responses  for each participant from the cleaned data}
\usage{
load_text_responses()
}
\value{
A data frame containing the only text responses from the cleaned data set
}
\description{
Text responses  for each participant from the cleaned data
}
