% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoreyHoekstra2019-package.R
\docType{package}
\name{MoreyHoekstra2019}
\alias{MoreyHoekstra2019}
\title{MoreyHoekstra2019: A package containing data, code, and demos for Morey & Hoekstra 2019}
\section{MoreyHoekstra2019 functions}{

The MoreyHoekstra2019 functions ...
}

