% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_options.R
\name{pkg_options}
\alias{pkg_options}
\title{Set or get options for my package}
\usage{
pkg_options(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\description{
Set or get options for my package
}
\section{Supported options}{

The following options are supported
\itemize{
\item{\code{ggplot_family}}{(\code{character};"Lato Light") Font family to use for ggplot2 graphs }
\item{\code{base_family}}{(\code{character};"lato") Font family to use for base graphs }
}
}

