% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/santaDisplay_widget.R
\name{santaDisplay}
\alias{santaDisplay}
\title{HTML widget for displaying an animation of participant behavior}
\usage{
santaDisplay(value, label = "", tooltip = "", width = NULL,
  height = NULL, elementId = NULL)
}
\arguments{
\item{value}{Participant ID to display}

\item{label}{label for the widget}

\item{tooltip}{tooltip for the widget}

\item{width}{width of div}

\item{height}{height of div}

\item{elementId}{widget div element id}
}
\value{
An object of class \code{htmlwidget} for printing into HTML
}
\description{
HTML widget for displaying an animation of participant behavior
}
