% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_error_rates.R
\name{sigdet_model_fit}
\alias{sigdet_model_fit}
\title{Fit signal detection model to data}
\usage{
sigdet_model_fit(freqs)
}
\arguments{
\item{freqs}{summary data}
}
\value{

}
\description{
Fit signal detection model to data
}
