% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_demo.R
\name{task_demo}
\alias{task_demo}
\title{Start a web server with a demo of the experimental task and open browser}
\usage{
task_demo(toy_name = NULL, evidence_power = NULL)
}
\arguments{
\item{toy_name}{Name of the toy; determines effect size}

\item{evidence_power}{Power for the p value transform}
}
\value{
Invisibly returns a list of the job number of the HTTP server and url
}
\description{
Start a web server with a demo of the experimental task and open browser
}
