% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/christmas_stats_public-data.R
\docType{data}
\name{text_coding}
\alias{text_coding}
\title{Coding of the text responses}
\description{
Categorization of each participants' strategy-related
free-text responses
}
\details{
\itemize{
\item id: participant id
\item text_comparison: Response mentioned comparison to the random shuffle reports
\item text_asymmetry: Response mentioned symmetry/asymmetry in the experimental samples
\item text_sampling_var: Response mentioned random shuffle reports as a means to assess sampling variability,
distribution under the null, chance distribution, etc
\item text_inc_asymmetry: Response mentioned increasing asymmetry (or lack thereof) as sample size increased
\item text_no_shuffle: Explicitly said they did not use the shuffle reports
\item text_irrelevant: Text was irrelevant
\item text_missing: Text was missing
}
}
\author{
Richard D. Morey \email{richarddmorey@gmail.com}
}
\keyword{data}
